cli: Port Busybox 2021 hush to U-Boot.
This commit adds new file cli_hush_2021.c, it is a copy of Busybox hush file as
it was of time to commit 37460f5da.
It was modified to not compile some part specific to Busybox and adds some code
needed by U-Boot.
The modifications consists only on adding code #if(n)def guard like the
following:
// Busybox code
// U-Boot code

For the moment, this refurbished flavor of hush only permits running command
without any keywords (i.e., if and for are not recognized) or variable expansion
(i.e., echo $foo prints foo and not value stored in variable foo).

Signed-off-by: Francis Laniel <francis.laniel@amarulasolutions.com>